chore: add LICENSE file
Add a LICENSE file to the project to specify the terms and conditions for using the software. The license chosen is the MIT License, which grants permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software, with the condition that the copyright notice and permission notice are included in all copies or substantial portions of the software. The software is provided "as is" without warranty.