chore(package.json): add "@npmcli/map-workspaces" and "@types/npmcli__map-workspaces" dependencies
The "@npmcli/map-workspaces" package is added as a dependency to enable mapping of workspaces in the project. This package is required for managing multiple packages within a monorepo.

The "@types/npmcli__map-workspaces" package is added as a devDependency to provide TypeScript type definitions for the "@npmcli/map-workspaces" package. This helps with type checking and editor autocompletion when using the package in TypeScript files.