feat: add FlatDependenciesPackage type and buildFromDependencyGraph function
A new file `src/lib/gha/flat-dependencies-package.ts` was added. This file introduces a new type `FlatDependenciesPackage` which represents a package with flat dependencies. It has properties `kind`, `name`, and `dependencies`. The `buildFromDependencyGraph` function is also added, which takes a graph and a rootId and returns a `FlatDependenciesPackage` object. Currently, the function throws an error and needs to be implemented.

test: add convertToMap function test

In the file `src/lib/graph/digraph.test.ts`, a new test case is added for the `convertToMap` function. The test checks if the function correctly converts a graph to a Map object. The test case expects the result to be `Map([[0, [1]], [1, [2]], [2, []]])`.

feat: add convertToMap function

In the file `src/lib/graph/digraph.ts`, the `convertToMap` function is added. This function takes a graph and converts it to a Map object. It iterates over the graph and creates a new Map object with the id as the key and the dependencies as the value.